#17 wip: adjust some class names and returns
