+++ src/main/java/dev/orion/users/infra/repository/UserRepository.java
@@ -16,119 +16,42 @@
  */
 package dev.orion.users.infra.repository;
 
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.stream.Stream;
-
-import javax.enterprise.context.ApplicationScoped;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-
 import dev.orion.users.domain.model.User;
 import dev.orion.users.validation.dto.UserQuery;
-import io.quarkus.hibernate.reactive.panache.Panache;
-import io.quarkus.panache.common.Parameters;
+import io.quarkus.hibernate.reactive.panache.PanacheRepository;
+import io.smallrye.mutiny.Multi;
 import io.smallrye.mutiny.Uni;
 
+import java.util.List;
+
 /**
- * Implements the repository pattern for the user entity.
+ * User repository.
  */
-@ApplicationScoped
-public class UserRepository implements Repository {
+public interface UserRepository extends PanacheRepository<User> {
 
-  /**
-   * Creates a user in the service.
-   *
-   * @param name     : A name of the user
-   * @param email    : A valid e-mail
-   * @param password : A password of the user
-   *
-   * @return Returns a user asynchronously
-   */
-  @Override
-  public Uni<User> createUser(final String name, final String email,
-      final String password) {
+    /**
+     * Creates a user in the service.
+     *
+     * @param name     : A name of the user
+     * @param email    : A valid e-mail
+     * @param password : A password of the user
+     *
+     * @return Returns a user asynchronously
+     */
+    Uni<User> createUser(String name, String email, String password);
 
-    return checkEmail(email)
-        .onItem()
-        .ifNotNull()
-        .failWith(new IllegalArgumentException("The e-mail already exists"))
-        .onItem()
-        .ifNull()
-        .switchTo(() -> persistUser(name, email, password));
-  }
+    /**
+     * Returns a user looking for email and password.
+     *
+     * @param email    : An e-mail of the user
+     * @param password : A password
+     *
+     * @return Returns a user asynchronously
+     */
+    Uni<User> authenticate(String email, String password);
 
-  /**
-   * Verifies if the e-mail already exists in the database.
-   *
-   * @param email : An e-mail address
-   *
-   * @return Returns true if the e-mail already exists
-   */
-  private Uni<User> checkEmail(final String email) {
-    return find("email", email).firstResult();
-  }
+    Multi<User> listByQuery(UserQuery query);
 
-  /**
-   * Persists a user in the service.
-   *
-   * @param name     : The name of the user
-   * @param email    : An e-mail address of the a user
-   * @param password : The password of the user
-   *
-   * @return Returns Uni<User> object
-   */
-  private Uni<User> persistUser(final String name, final String email,
-      final String password) {
-    User user = new User();
-    user.setName(name);
-    user.setEmail(email);
-    user.setPassword(password);
-    return Panache.<User>withTransaction(user::persist);
-  }
-
-  /**
-   * Returns a user looking for email and password.
-   *
-   * @param email    : An e-mail of the user
-   * @param password : A password
-   *
-   * @return Returns a user asynchronously
-   */
-  @Override
-  public Uni<User> authenticate(final String email, final String password) {
-    Map<String, Object> params = Parameters.with("email", email)
-        .and("password", password).map();
-    return find("email = :email and password = :password", params)
-        .firstResult();
-  }
-
-  /**
-   * @param query
-   * @return
-   */
-  @Override
-  public Uni<List<User>> listByQuery(UserQuery query) {
-    ObjectMapper mapper = new ObjectMapper();
-    Map<String, Object> params = mapper.convertValue(query, Map.class);
-
-    if (params.values().stream().allMatch(Objects::isNull)) {
-      return findAll().list();
-    }
-
-    return list("hash = :hash or name like :name or email = :email", params);
-
-  }
-
-  /**
-   * @param id
-   * @return
-   */
-  @Override
-  public Uni<Long> removeUser(String id) {
-
-    return delete("hash", id);
-  }
+    Uni<Long> removeUser(String id);
 
 }
