+++ src/main/java/dev/orion/users/usecase/CreateUser.java
@@ -6,8 +6,8 @@ import org.apache.commons.codec.digest.DigestUtils;
 import org.apache.commons.validator.routines.EmailValidator;
 
 import dev.orion.users.domain.model.User;
-import dev.orion.users.infra.repository.Repository;
 import dev.orion.users.infra.repository.UserRepository;
+import dev.orion.users.infra.repository.UserRepositoryImpl;
 import io.smallrye.mutiny.Uni;
 
 @ApplicationScoped
@@ -15,7 +15,7 @@ public class CreateUser implements UseCase {
     private static final int SIZE_PASSWORD = 8;
 
     /** User repository. */
-    private Repository repository = new UserRepository();
+    private UserRepository repository = new UserRepositoryImpl();
 
     /**
      * Creates a user in the service (UC: Create).
