+++ src/main/java/dev/orion/users/usecase/ListUser.java
@@ -1,22 +1,20 @@
 package dev.orion.users.usecase;
 
 import dev.orion.users.domain.model.User;
-import dev.orion.users.infra.repository.Repository;
 import dev.orion.users.infra.repository.UserRepository;
+import dev.orion.users.infra.repository.UserRepositoryImpl;
 import dev.orion.users.validation.dto.UserQuery;
-import io.smallrye.mutiny.Uni;
-
-import java.util.List;
+import io.smallrye.mutiny.Multi;
 
 public class ListUser implements UseCase {
 
-    private Repository repository = new UserRepository();
+    private UserRepository repository = new UserRepositoryImpl();
 
     /**
      * @return
      */
     @Override
-    public Uni<List<User>> listUser(UserQuery query) {
+    public Multi<User> listUser(UserQuery query) {
         return repository.listByQuery(query);
     }
 }
