+++ src/main/java/dev/orion/users/usecase/RemoveUser.java
@@ -1,13 +1,12 @@
 package dev.orion.users.usecase;
 
-import dev.orion.users.domain.model.User;
-import dev.orion.users.infra.repository.Repository;
 import dev.orion.users.infra.repository.UserRepository;
+import dev.orion.users.infra.repository.UserRepositoryImpl;
 import io.smallrye.mutiny.Uni;
 
 public class RemoveUser implements UseCase {
 
-    private Repository repository = new UserRepository();
+    private UserRepository repository = new UserRepositoryImpl();
 
     @Override
     public Uni<Long> removeUser(String hash) {
