+++ src/test/java/dev/orion/users/usecase/ListUserTest.java
@@ -1,16 +1,14 @@
 package dev.orion.users.usecase;
 
 import dev.orion.users.domain.model.User;
-import dev.orion.users.infra.repository.Repository;
+import dev.orion.users.infra.repository.UserRepository;
 import dev.orion.users.validation.dto.UserQuery;
 import io.quarkus.test.junit.mockito.InjectSpy;
-import io.smallrye.mutiny.Uni;
+import io.smallrye.mutiny.Multi;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.mockito.Mockito.spy;
 
-import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Order;
@@ -18,7 +16,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestMethodOrder;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
-import org.mockito.*;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
@@ -29,29 +26,28 @@ public class ListUserTest {
 
     @InjectSpy
     @Mock
-    Repository repository;
+    UserRepository repository;
 
     @InjectSpy
     @InjectMocks
-    ListUser listUsersUC = new ListUser();
+    UseCase useCase = new ListUser();
 
     @Test
     @DisplayName("List All Users")
     @Order(1)
     void listAllUsersTest() {
         UserQuery query = new UserQuery();
-        query.setEmail(null);
-        query.setHash(null);
-        query.setName(null);
-        Uni<User> user = Uni.createFrom().item(new User());
-        Uni<List<User>> userList = Uni.join().all(user).andCollectFailures();
 
+        Multi<User> userList = Multi.createFrom().item(createUserMock("123", "Orion", "Orion@email.com"));
         Mockito.when(repository.listByQuery(query)).thenReturn(userList);
 
-        Uni<List<User>> users = listUsersUC.listUser(query);
+        Multi<User> users = useCase.listUser(query);
+
         Mockito.verify(repository, Mockito.times(1)).listByQuery(query);
 
-        assertEquals(userList, users);
+        assertNotNull(query);
+        assertNotNull(users);
+        assertEquals(users, userList);
     }
 
     @Test
@@ -59,18 +55,14 @@ public class ListUserTest {
     @Order(2)
     void listUsersByNameTest() {
         UserQuery query = new UserQuery();
-        query.setName("Test");
+        query.setName("Orion");
 
-        Uni<User> user = Uni.createFrom().item(this.createUserMock("Teste", "Teste", "teste@email.com"));
-
-        Uni<List<User>> userList = Uni.join().all(user).andCollectFailures();
-
+        Multi<User> userList = Multi.createFrom().item(createUserMock("123", "Orion", "Orion@email.com"));
         Mockito.when(repository.listByQuery(query)).thenReturn(userList);
 
-        Uni<List<User>> users = listUsersUC.listUser(query);
+        Multi<User> users = useCase.listUser(query);
 
         Mockito.verify(repository, Mockito.times(1)).listByQuery(query);
-
         assertNotNull(userList);
         assertNotNull(users);
         assertEquals(users, userList);
@@ -81,21 +73,18 @@ public class ListUserTest {
     @Order(3)
     void listUsersByIdTest() {
         UserQuery query = new UserQuery();
-        query.setHash("Teste");
+        query.setHash("123");
 
-        Uni<User> user = Uni.createFrom().item(this.createUserMock("Teste", "Teste", "teste@email.com"));
-
-        Uni<List<User>> userList = Uni.join().all(user).andCollectFailures();
-
+        Multi<User> userList = Multi.createFrom().item(createUserMock("123", "Orion", "Orion@email.com"));
         Mockito.when(repository.listByQuery(query)).thenReturn(userList);
 
-        Uni<List<User>> users = listUsersUC.listUser(query);
+        Multi<User> users = useCase.listUser(query);
 
         Mockito.verify(repository, Mockito.times(1)).listByQuery(query);
 
         assertNotNull(userList);
         assertNotNull(users);
-        assertEquals(users, userList);
+        assertEquals(userList, users);
     }
 
     private User createUserMock(String hash, String name, String email) {
